Adding support for python objects in SQL RAS by using SQLA.PickleType
